Rename LicenseMetaData to License (and drop previous, duplicate, License)
